More postgres/db conf and moving to com.marcusfromsweden.plantdoctor as top package
