+++ pom.xml
@@ -21,7 +21,7 @@
         <spring-boot.version>3.1.7</spring-boot.version>
         <maven-compiler-plugin.version>3.11.0</maven-compiler-plugin.version>
         <maven-surefire-and-failsafe-plugin.version>3.1.2</maven-surefire-and-failsafe-plugin.version>
-        <json-path.version>2.7.0</json-path.version>
+        <json-path.version>2.9.0</json-path.version>
         <logback.version>1.5.6</logback.version>
         <springdoc-openapi.version>2.1.0</springdoc-openapi.version>
         <maven-dependency-plugin.version>3.6.0</maven-dependency-plugin.version>
@@ -67,8 +67,8 @@
         </dependency>
 
         <dependency>
-            <groupId>com.h2database</groupId>
-            <artifactId>h2</artifactId>
+            <groupId>org.postgresql</groupId>
+            <artifactId>postgresql</artifactId>
             <scope>runtime</scope>
         </dependency>
