+++ src/main/java/com/marcusfromsweden/plantdoctor/PlantDoctorApplication.java
@@ -1,13 +1,13 @@
-package com.marcusfromsweden.plantdoctorh2;
+package com.marcusfromsweden.plantdoctor;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class PlantDoctorH2Application {
+public class PlantDoctorApplication {
 
     public static void main(String[] args) {
-        SpringApplication.run(PlantDoctorH2Application.class, args);
+        SpringApplication.run(PlantDoctorApplication.class, args);
     }
 
 }
