+++ src/main/java/com/marcusfromsweden/plantdoctor/controller/PlantController.java
@@ -1,4 +1,4 @@
-package com.marcusfromsweden.plantdoctorh2.controller;
+package com.marcusfromsweden.plantdoctor.controller;
 
 import java.util.List;
 import java.util.Optional;
@@ -15,8 +15,8 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import com.marcusfromsweden.plantdoctorh2.entity.Plant;
-import com.marcusfromsweden.plantdoctorh2.service.PlantService;
+import com.marcusfromsweden.plantdoctor.entity.Plant;
+import com.marcusfromsweden.plantdoctor.service.PlantService;
 import jakarta.validation.Valid;
 
 @RestController
