+++ src/main/java/com/marcusfromsweden/plantdoctor/entity/Plant.java
@@ -1,4 +1,4 @@
-package com.marcusfromsweden.plantdoctorh2.entity;
+package com.marcusfromsweden.plantdoctor.entity;
 
 import java.time.LocalDate;
 import jakarta.persistence.Entity;
