+++ src/main/java/com/marcusfromsweden/plantdoctor/repository/PlantRepository.java
@@ -1,8 +1,8 @@
-package com.marcusfromsweden.plantdoctorh2.repository;
+package com.marcusfromsweden.plantdoctor.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import com.marcusfromsweden.plantdoctorh2.entity.Plant;
+import com.marcusfromsweden.plantdoctor.entity.Plant;
 
 @Repository
 public interface PlantRepository extends JpaRepository<Plant, Long> {
