+++ src/main/resources/application.properties
@@ -1,24 +1,40 @@
-spring.application.name=plant-doctor-h2
-#spring.datasource.url=${SPRING_DATASOURCE_PLANTDOCTOR_URL:jdbc:postgresql://localhost:5432/plantdoctor}
-#spring.datasource.username=${SPRING_DATASOURCE_PLANTDOCTOR_USERNAME:plantdoctoruser}
-#spring.datasource.password=${SPRING_DATASOURCE_PLANTDOCTOR_PASSWORD:plantdoctorpassword}
-server.port=8080
+spring.application.name=plant-doctor-postgres
+spring.datasource.url=${SPRING_DATASOURCE_PLANTDOCTOR_URL:jdbc:postgresql://localhost:5432/plantdoctor}
+spring.datasource.username=${SPRING_DATASOURCE_PLANTDOCTOR_USERNAME:plantdoctoradmin}
+spring.datasource.password=${SPRING_DATASOURCE_PLANTDOCTOR_PASSWORD:plantdoctoradminpassword}
+spring.jpa.properties.hibernate.default_schema=study
 
 # Set the logging level
 logging.level.root=INFO
 logging.level.com.example.plantmanager=DEBUG
 
 # Specify the log file name and location
-logging.file.name=logs/plant-doctor-h2.log
+logging.file.name=logs/plant-doctor-postgres.log
+
+spring.jackson.time-zone=Europe/Stockholm
+spring.web.locale=sv_SE
+spring.web.locale-resolver=fixed
+
+# Hibernate (JPA) Configuration
+spring.jpa.database-platform=org.hibernate.dialect.PostgreSQLDialect
+spring.jpa.hibernate.ddl-auto=update
+spring.jpa.show-sql=true
+spring.jpa.properties.hibernate.format_sql=true
+# Disable open-in-view to avoid performance issues
+spring.jpa.open-in-view=false
 
 # Configure log file rotation
 #logging.logback.rollingpolicy.file-name-pattern=logs/plantdoctor-%d{yyyy-MM-dd}.%i.log
 #logging.logback.rollingpolicy.max-file-size=10MB
 #logging.logback.rollingpolicy.max-history=30
 
-spring.jackson.time-zone=Europe/Stockholm
-spring.web.locale=sv_SE
-spring.web.locale-resolver=fixed
+# Hibernate (JPA) Configuration for generating SQL scripts
+## use below to generate the SQL using hbm2ddl
+#spring.jpa.hibernate.ddl-auto=none
+#spring.jpa.properties.hibernate.hbm2ddl.schema-generation.scripts.action=create
+#spring.jpa.properties.hibernate.hbm2ddl.schema-generation.scripts.create-target=src/main/resources/schema.sql
+#spring.jpa.properties.hibernate.hbm2ddl.schema-generation.scripts.drop-target=src/main/resources/drop.sql
+#spring.jpa.properties.hibernate.hbm2ddl.schema-generation.create-source=metadata
+#spring.jpa.properties.hibernate.hbm2ddl.schema-generation.drop-source=metadata
 
-# Disable open-in-view to avoid performance issues
-spring.jpa.open-in-view=false
+server.port=8080
