+++ src/test/java/com/marcusfromsweden/plantdoctor/PlantDoctorApplicationTests.java
@@ -1,10 +1,10 @@
-package com.marcusfromsweden.plantdoctorh2;
+package com.marcusfromsweden.plantdoctor;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
-@SpringBootTest
-class PlantDoctorH2ApplicationTests {
+@SpringBootTest(classes = PlantDoctorApplication.class)
+class PlantDoctorApplicationTests {
 
     @Test
     void contextLoads() {
